separate file for matrix error
